+++ crates/simulate/src/simulate.rs
@@ -19,11 +19,9 @@ use cairo_vm::vm::runners::cairo_runner::ExecutionResources;
 use cheatnet::runtime_extensions::call_to_blockifier_runtime_extension::execution::entry_point::execute_call_entry_point;
 use cheatnet::runtime_extensions::call_to_blockifier_runtime_extension::rpc::CallFailure;
 use cheatnet::runtime_extensions::call_to_blockifier_runtime_extension::rpc::CallResult;
-use cheatnet::runtime_extensions::forge_runtime_extension::cheatcodes::spy_events::Event;
 use cheatnet::state::CheatnetState;
 use internal_tracing::build_debugger_data::debugger_data_maps_full_class_to_class;
 use internal_tracing::debugger_data_fetcher::fetch_classes_debugger_data;
-use internal_tracing::event_calls_map;
 use internal_tracing::event_calls_map::EventCallsMap;
 use internal_tracing::SimulationDebuggerData;
 use num_traits::ToPrimitive;
@@ -47,6 +45,9 @@ use url::Url;
 use walnut_shared::felt_to_field_element;
 use walnut_shared::felts_to_string;
 use walnut_shared::field_element_to_felt;
+use walnut_shared::EnumAbi;
+use walnut_shared::EventAbi;
+use walnut_shared::StructAbi;
 use walnut_shared::{chain_id_to_readable_string, create_rpc_client_from_url};
 
 use crate::abi_processor::AbiProcessor;
@@ -54,6 +55,7 @@ use crate::contract_calls_map::ContractCallsMap;
 use crate::contract_calls_map::ContractCallsMapBuilder;
 use crate::contract_names::ContractNamesFetcher;
 use crate::debugger_trace::DebuggerTraceBuilder;
+use crate::events::EmittedEvent;
 use crate::function_calls::create_function_calls_map;
 use crate::state::ForkStateReader;
 use crate::transaction_extraction::extract_block_number_transaction_receipt;
@@ -61,6 +63,7 @@ use crate::transaction_extraction::extract_block_timestamp;
 use crate::transaction_extraction::extract_block_txs_info;
 use crate::transaction_extraction::extract_chain_id_from_felt;
 use crate::transaction_extraction::extract_execution_status_transaction_receipt;
+use crate::transaction_extraction::extract_starkgate_event_transaction_receipt;
 use crate::transaction_extraction::extract_submitted_tx;
 use crate::transaction_extraction::extract_transaction_contex;
 use crate::utils::calldata_to_hex;
@@ -85,6 +88,7 @@ pub async fn simulate(
     } else {
         provider_client.block_number().await?
     };
+
     let (block_info, transaction_index, tx_number_in_block) =
         extract_block_txs_info(&provider_client, &args, block_number).await?;
 
@@ -95,11 +99,14 @@ pub async fn simulate(
             block_number,
             transaction_index,
             tx_number_in_block,
+            db_pool,
+            s3_client,
         )
         .map_err(|e| {
             TransactionSimulationError::StateError(StateError::StateReadError(e.to_string()))
         })?,
     );
+    let strkgate_event = args.strkgate_event.clone();
 
     let cheatnet_state = run_simulation(block_info, args, &mut cached_fork_state)?;
 
@@ -107,6 +114,7 @@ pub async fn simulate(
         mut contract_calls_map,
         mut next_call_id,
         deepest_failed_contract_call_id,
+        cheatnet_state_detected_events,
         ..
     } = ContractCallsMapBuilder::new_from_cheatnet_state(cheatnet_state);
 
@@ -122,6 +130,10 @@ pub async fn simulate(
     );
 
     filter_and_hide_unlinked_function_calls(&mut contract_calls_map, &function_calls_map);
+
+    let mut event_abis: Vec<EventAbi> = Vec::new();
+    let mut struct_abis: Vec<StructAbi> = Vec::new();
+    let mut enum_abis: Vec<EnumAbi> = Vec::new();
 
     for call in contract_calls_map.0.values_mut() {
         if let Some(class_hash) = call.entry_point.class_hash {
@@ -147,14 +159,41 @@ pub async fn simulate(
                 call.decode_call_result(&abi_processor.struct_abis, &abi_processor.enum_abis);
                 call.decode_call_arguments(&abi_processor.struct_abis, &abi_processor.enum_abis);
 
-                //event_abis.extend(abi_processor.event_abis);
-                //struct_abis.extend(abi_processor.struct_abis);
-                //enum_abis.extend(abi_processor.enum_abis);
+                event_abis.extend(abi_processor.event_abis);
+                struct_abis.extend(abi_processor.struct_abis);
+                enum_abis.extend(abi_processor.enum_abis);
             }
         }
     }
 
-    ContractNamesFetcher::new(provider_client, &chain_id)
+    let mut contract_names_fetcher = ContractNamesFetcher::new(provider_client, &chain_id);
+    // The StarkNet transaction emits this `Transfer` event from the StarkGate ETH token contract.
+    // This event is present inside the transaction receipt but is not found in the Foundry-emitted
+    // events array.
+    // To maintain consistency with blockchain explorers, we need to manually append this event
+    // to the vector of all events.
+    let strkgate_emitted_event = if let Some(event) = strkgate_event {
+        let contract_address_felt = field_element_to_felt(event.from_address);
+        let contract_address_str = contract_address_felt.to_fixed_hex_string();
+
+        let contract_name = contract_names_fetcher
+            .fetch_single_contract_name(contract_address_str)
+            .await;
+        EmittedEvent::convert_event_to_emitted_event(&event, &contract_name)
+    } else {
+        None
+    };
+
+    let events = EmittedEvent::create_emitted_events_list(
+        &mut contract_calls_map,
+        &event_abis,
+        &struct_abis,
+        &enum_abis,
+        &cheatnet_state_detected_events,
+        strkgate_emitted_event,
+    );
+
+    contract_names_fetcher
         .set_contract_names(&mut contract_calls_map)
         .await;
 
@@ -178,6 +217,7 @@ pub async fn simulate(
         contract_calls_map,
         function_calls_map,
         event_calls_map,
+        events,
         execution_result,
         simulation_debugger_data: Some(SimulationDebuggerData {
             classes_debugger_data: debugger_data_maps_full_class_to_class(classes_debugger_data),
@@ -365,6 +405,7 @@ pub async fn simulate_transaction_by_hash(
                                 contract_calls_map: ContractCallsMap::new(),
                                 function_calls_map: FunctionCallsMap::new(),
                                 event_calls_map: EventCallsMap::default(),
+                                events: Vec::new(),
                                 execution_result: ExecutionResult::Reverted { reason },
                                 simulation_debugger_data: Some(SimulationDebuggerData {
                                     classes_debugger_data: HashMap::new(),
@@ -387,6 +428,8 @@ pub async fn simulate_transaction_by_hash(
                         }
                     }
 
+                    let strkgate_event =
+                        extract_starkgate_event_transaction_receipt(&transaction_receipt);
                     // Perform transaction simulation
                     let (simulation_result, block_timestamp, transaction_index_in_block) =
                         simulate(
@@ -405,6 +448,7 @@ pub async fn simulate_transaction_by_hash(
                                 transaction_type: Some(transaction_type),
                                 resource_bounds: Some(resource_bounds),
                                 paymaster_data: Some(paymaster_data),
+                                strkgate_event,
                             },
                         )
                         .await?;