+++ crates/verification/src/helpers.rs
@@ -137,25 +137,22 @@ pub async fn process_new_cairo_version_verification(
     verification_id: Uuid,
     class_verification_data: &mut ClassVerificationData,
 ) -> Result<()> {
-    let mut classes_to_verify_map: HashMap<String, ContractClass> = HashMap::new();
-    let encountered_error = spawn_new_cairo_version_verification_tasks(
+    let mut classes_to_verify_map: HashMap<String, (ContractClass, String)> = HashMap::new();
+    if let Err(encountered_error) = spawn_new_cairo_version_verification_tasks(
         manifest,
         tmp_dir,
         db_pool,
         verification_id,
         &mut classes_to_verify_map,
     )
-    .await;
-
-    if encountered_error {
-        if let Err(move_err) = move_failed_verification_to_failed_tmp(tmp_dir) {
+    .await
+    {
+        if let Err(move_err) = move_failed_verification_to_failed_tmp(tmp_dir, &verification_id) {
             let err = format!("Failed to move verification to failed tmp: {:?}", move_err);
             error!("{:?}", err);
             return Err(anyhow::anyhow!(err));
         }
-        return Err(anyhow::anyhow!(
-            "Verification failed due to build project errors."
-        ));
+        return Err(encountered_error);
     }
 
     update_new_cairo_version_class_verification_data(
@@ -171,14 +168,50 @@ async fn spawn_new_cairo_version_verification_tasks(
     tmp_dir: &PathBuf,
     db_pool: &Pool<Postgres>,
     verification_id: Uuid,
-    classes_to_verify_map: &mut HashMap<String, ContractClass>,
-) -> bool {
+    classes_to_verify_map: &mut HashMap<String, (ContractClass, String)>,
+) -> Result<(), anyhow::Error> {
     // Broadcast channel for error signalization
     let (tx, _) = tokio::sync::broadcast::channel(1);
+    let mut encountered_error: Option<anyhow::Error> = None;
+    let mut inline_class_hashes: Vec<(String, ContractClass)> = Vec::new();
 
+    // First build profil with inline strategy, if it exists
+    if let Some(inline_strategy_profile) = manifest.profile_with_inline_strategy.keys().next() {
+        match build_with_scarb_for_profile(manifest, tmp_dir, inline_strategy_profile) {
+            Ok(classes) => {
+                for (class_hash, contract_class) in classes {
+                    inline_class_hashes.push((class_hash.clone(), contract_class.clone()));
+                    if let Err(err) = insert_class_hash_profiles(
+                        db_pool,
+                        &class_hash,
+                        inline_strategy_profile,
+                        verification_id,
+                        &true,
+                        Some(&class_hash),
+                    )
+                    .await
+                    {
+                        error!(
+                            "Failed to insert inline strategy class hash profile: {:?}",
+                            err
+                        );
+                    }
+                    classes_to_verify_map
+                        .insert(class_hash.clone(), (contract_class, class_hash.clone()));
+                }
+            }
+            Err(e) => {
+                error!("Failed to build inline strategy profile: {:?}", e);
+                encountered_error = Some(e);
+            }
+        }
+    }
+
+    // Buld other profiles, skip inline profile
     let mut futures: FuturesUnordered<_> = manifest
         .profiles
         .iter()
+        .filter(|&profile| !manifest.profile_with_inline_strategy.contains_key(profile))
         .cloned()
         .map(|profile| {
             let tmp_dir_clone = tmp_dir.clone();
@@ -196,59 +229,78 @@ async fn spawn_new_cairo_version_verification_tasks(
                         Err(anyhow::anyhow!("Build project failed."))
                     }
                 result = async {
-                match build_with_scarb_for_profile(&manifest_clone, &tmp_dir_clone, &profile) {
-                    Ok(classes) => Ok::<_, anyhow::Error>((classes, profile.clone())),
-                    Err(e) => {
-                        // Send signal to other thread to terminate
-                        let _ = tx.send(());
-                        Err(e)
+                    match build_with_scarb_for_profile(&manifest_clone, &tmp_dir_clone, &profile) {
+                        Ok(classes) => Ok::<_, anyhow::Error>((classes, profile.clone())),
+                        Err(e) => {
+                            // Send signal to other thread to terminate
+                            let _ = tx.send(());
+                            Err(e)
+                        }
                     }
-                }
                         } => result
                 }
             })
         })
         .collect();
 
-    let mut encountered_error = false;
-
     while let Some(result) = futures.next().await {
         match result {
             Ok(Ok((classes, profile))) => {
-                for (class_hash, contract_class) in classes {
-                    if let Err(err) =
-                        insert_class_hash_profiles(db_pool, &class_hash, &profile, verification_id)
+                if classes.len() == inline_class_hashes.len() {
+                    for (idx, (class_hash, _contract_class)) in classes.into_iter().enumerate() {
+                        if let Some((inline_class_hash, inline_contract_class)) =
+                            inline_class_hashes.get(idx).cloned()
+                        {
+                            if let Err(err) = insert_class_hash_profiles(
+                                db_pool,
+                                &class_hash,
+                                &profile,
+                                verification_id,
+                                &false,
+                                Some(&inline_class_hash),
+                            )
                             .await
-                    {
-                        error!("Failed to insert class hash with profile: {:?}", err);
-                    }
+                            {
+                                error!("Failed to insert class hash with profile: {:?}", err);
+                            }
 
-                    classes_to_verify_map
-                        .entry(class_hash)
-                        .or_insert(contract_class);
+                            classes_to_verify_map
+                                .entry(class_hash)
+                                .or_insert((inline_contract_class, inline_class_hash));
+                        }
+                    }
                 }
             }
-            Ok(Err(_e)) => {
-                encountered_error = true;
+            Ok(Err(e)) => {
+                encountered_error = Some(e);
             }
             Err(e) => {
                 error!("Tokio task failed: {:?}", e);
-                encountered_error = true;
+                encountered_error = Some(e.into());
             }
         }
     }
 
-    encountered_error
+    if let Some(e) = encountered_error {
+        return Err(e);
+    }
+
+    Ok(())
 }
 
 fn update_new_cairo_version_class_verification_data(
     class_verification_data: &mut ClassVerificationData,
-    classes_to_verify_map: &HashMap<String, ContractClass>,
+    classes_to_verify_map: &HashMap<String, (ContractClass, String)>,
 ) {
     for (class_hash, class_result) in class_verification_data.iter_mut() {
-        if let Some(class) = classes_to_verify_map.get(class_hash) {
-            if let Ok((_, _, _, ref mut contract_class, _, _)) = class_result {
-                *contract_class = Some(class.clone());
+        if let Some((inline_contract_class, inline_class_hash)) =
+            classes_to_verify_map.get(class_hash)
+        {
+            if let Ok((_, _, _, ref mut contract_class, ref mut inline_strategy_class_hash, _, _)) =
+                class_result
+            {
+                *contract_class = Some(inline_contract_class.clone());
+                *inline_strategy_class_hash = Some(inline_class_hash.clone());
             }
         } else {
             let error_message = format!(
@@ -270,9 +322,9 @@ pub async fn process_old_cairo_version_verification(
     verification_id: Uuid,
     class_verification_data: &mut ClassVerificationData,
 ) -> Result<()> {
-    let mut classes_to_verify_map = HashMap::new();
+    let mut classes_to_verify_map: HashMap<String, (ContractClass, PathBuf)> = HashMap::new();
 
-    let encountered_error = spawn_old_cairo_version_verification_tasks(
+    if let Err(encountered_error) = spawn_old_cairo_version_verification_tasks(
         manifest,
         cairo_version,
         tmp_dir,
@@ -280,17 +332,14 @@ pub async fn process_old_cairo_version_verification(
         verification_id,
         &mut classes_to_verify_map,
     )
-    .await;
-
-    if encountered_error {
-        if let Err(move_err) = move_failed_verification_to_failed_tmp(tmp_dir) {
+    .await
+    {
+        if let Err(move_err) = move_failed_verification_to_failed_tmp(tmp_dir, &verification_id) {
             let err = format!("Failed to move verification to failed tmp: {:?}", move_err);
             error!("{:?}", err);
             return Err(anyhow::anyhow!(err));
         }
-        return Err(anyhow::anyhow!(
-            "Verification failed due to project build errors."
-        ));
+        return Err(encountered_error);
     }
 
     update_old_cairo_version_class_verification_data(
@@ -308,10 +357,12 @@ async fn spawn_old_cairo_version_verification_tasks(
     db_pool: &Pool<Postgres>,
     verification_id: Uuid,
     classes_to_verify_map: &mut HashMap<String, (ContractClass, PathBuf)>,
-) -> bool {
+) -> Result<(), anyhow::Error> {
     // Broadcast channel for error signalization
     let (tx, _) = tokio::sync::broadcast::channel(1);
+    let mut encountered_error: Option<anyhow::Error> = None;
 
+    // Buld profiles, skip inline profile
     let mut futures: FuturesUnordered<_> = manifest
         .profiles
         .iter()
@@ -331,33 +382,38 @@ async fn spawn_old_cairo_version_verification_tasks(
                         Err(anyhow::anyhow!("Scarb build failed"))
                     }
                 // Compile with Scarb for the given profile
-                result = async {match compile_with_scarb_for_profile(
-                    &manifest_clone,
-                    cairo_version,
-                    &tmp_dir_clone,
-                    &profile,
-                ) {
-                    Ok(classes) => Ok((classes, profile)),
-                    Err(e) => {
-                        let _ = tx.send(());
-                        Err(e)
-                    }
-                }
-                        } => result
+                result = async {
+                    match compile_with_scarb_for_profile(
+                        &manifest_clone,
+                        cairo_version,
+                        &tmp_dir_clone,
+                        &profile,
+                        ) {
+                            Ok(classes) => Ok((classes, profile)),
+                            Err(e) => {
+                                let _ = tx.send(());
+                                Err(e)
+                            }
+                        }
+                    } => result
                 }
             })
         })
         .collect();
 
-    let mut encountered_error = false;
-
     while let Some(result) = futures.next().await {
         match result {
             Ok(Ok((classes, profile))) => {
                 for (class_hash, contract_class, cairo_debug_info_path) in classes {
-                    if let Err(err) =
-                        insert_class_hash_profiles(db_pool, &class_hash, &profile, verification_id)
-                            .await
+                    if let Err(err) = insert_class_hash_profiles(
+                        db_pool,
+                        &class_hash,
+                        &profile,
+                        verification_id,
+                        &false,
+                        None,
+                    )
+                    .await
                     {
                         error!("Failed to insert class hash with profile: {:?}", err);
                     }
@@ -367,17 +423,19 @@ async fn spawn_old_cairo_version_verification_tasks(
                         .or_insert((contract_class, cairo_debug_info_path));
                 }
             }
-            Ok(Err(_e)) => {
-                encountered_error = true;
-            }
+            Ok(Err(e)) => encountered_error = Some(e),
             Err(e) => {
                 error!("Tokio task failed: {:?}", e);
-                encountered_error = true;
+                encountered_error = Some(e.into())
             }
         }
     }
 
-    encountered_error
+    if let Some(e) = encountered_error {
+        return Err(e);
+    }
+
+    Ok(())
 }
 
 fn update_old_cairo_version_class_verification_data(
@@ -385,21 +443,23 @@ fn update_old_cairo_version_class_verification_data(
     classes_to_verify_map: &HashMap<String, (ContractClass, PathBuf)>,
 ) {
     for (class_hash, class_result) in class_verification_data.iter_mut() {
-        if let Some((class, cairo_debug_info_path)) = classes_to_verify_map.get(class_hash) {
+        if let Some((contract_class, cairo_debug_info_path)) = classes_to_verify_map.get(class_hash)
+        {
             // If the class is found, update the contract_class and cairo_info_path in class_result
             if let Ok((
                 _,
                 program_from_blockchain,
                 _,
                 ref mut existing_contract_class,
+                _,
                 ref mut existing_cairo_debug_info_path,
                 ref mut existing_cairo_debug_info,
             )) = class_result
             {
-                *existing_contract_class = Some(class.clone());
+                *existing_contract_class = Some(contract_class.clone());
                 *existing_cairo_debug_info_path = Some(cairo_debug_info_path.clone());
 
-                if !programs_match(class, program_from_blockchain) {
+                if !programs_match(contract_class, program_from_blockchain) {
                     let err = anyhow::anyhow!(
                         "Contract class does not match for class hash: {}",
                         class_hash
@@ -440,7 +500,7 @@ fn programs_match(class: &ContractClass, program_from_blockchain: &[Felt]) -> bo
         .all(|(e1, e2)| e1.value.to_string() == e2.to_string())
 }
 
-fn load_cairo_debug_info(cairo_debug_info_path: &PathBuf) -> Result<SierraToCairoDebugInfo> {
+pub fn load_cairo_debug_info(cairo_debug_info_path: &PathBuf) -> Result<SierraToCairoDebugInfo> {
     let cairo_debug_info_file = File::open(cairo_debug_info_path).map_err(|e| {
         anyhow::anyhow!(
             "Failed to open debug info file {}: {:?}",