+++ crates/verification/src/utils.rs
@@ -50,8 +50,8 @@ pub fn create_temp_directory() -> Result<PathBuf> {
 
 // Failed verifications data for further investigation.
 // There is no auto removal from this location.
-pub fn move_failed_verification_to_failed_tmp(tmp_dir: &PathBuf) -> Result<()> {
-    let failed_tmp_dir = PathBuf::from("tmp/failed-verification");
+pub fn move_failed_verification_to_failed_tmp(tmp_dir: &PathBuf, verification_id: &Uuid) -> Result<()> {
+    let failed_tmp_dir = PathBuf::from(format!("tmp/failed-verification/{}", verification_id));
 
     error!(
         "Failed to verify classes - moving {} to {} for further investigation.",
@@ -70,4 +70,25 @@ pub fn move_failed_verification_to_failed_tmp(tmp_dir: &PathBuf) -> Result<()> {
     fs::remove_dir_all(tmp_dir)?;
 
     Ok(())
+}
+
+pub fn remove_walnut_debug_from_scarb(source_code: &mut HashMap<String, String>) {
+    if let Some(scarb_toml_contents) = source_code.get("Scarb.toml") {
+        if let Ok(mut scarb_toml_parsed) = scarb_toml_contents.parse::<toml::Value>() {
+            if let Some(profile_table) = scarb_toml_parsed
+                .get_mut("profile")
+                .and_then(toml::Value::as_table_mut)
+            {
+                profile_table.remove("walnut-debug");
+                match toml::to_string(&scarb_toml_parsed) {
+                    Ok(updated_scarb_config) => {
+                        source_code.insert("Scarb.toml".to_string(), updated_scarb_config);
+                    }
+                    Err(e) => {
+                        error!("Failed to serialize and update Scarb.toml: {}", e);
+                    }
+                }
+            }
+        }
+    }
 }